+++ alluxio/posix/const.py
@@ -12,17 +12,5 @@ class Constants:
     ALLUXIO_FILESYSTEM_TYPE = 'alluxio'
     S3_FILESYSTEM_TYPE = 's3'
 
-    # # OSS config keys
-    # OSS_ACCESS_KEY_ID = 'access_key_id'
-    # OSS_ACCESS_KEY_SECRET = 'access_key_secret'
-    # OSS_ENDPOINT = 'endpoint'
-    #
-    # # Alluxio config keys
-    # ALLUXIO_ETCD_ENABLE = "alluxio_etcd_enable"
-    # ALLUXIO_ETCD_HOST = 'alluxio_etcd_host'
-    # ALLUXIO_WORKER_HOSTS = 'alluxio_worker_hosts'
-    # ALLUXIO_BACKUP_FS = 'alluxio_backup_fs'
-    # ALLUXIO_ENABLE = 'alluxio_enable'
-
     # assist constants
     ALLUXIO_SEP_SIGN = '_'
